Add reason message to OuterLoopAttribute
We're often using comments to indicate why a test is being marked as OuterLoop.  This lets us do that as a string.  Then once all use has been updated to specify a message, we can remove the ctor without the message, and devs will be forced to specify a reason for why a test is marked as OuterLoop.