Handle non-unique IDs in VSCT files
In a VSCT file elements are uniquely identified by a combination of GUID
and ID. However, currently XliffTasks only considers the ID and the tool
will crash if it encounters the same ID reused under multiple GUIDs.

To fix this we need to incorporate the GUID into the unique identifiers
we generate for every localizable resource. However, in repos that are
already using XliffTasks we don't want to re-label all existing
localizable items in VSCT files as that will cause a good bit of churn
and busy work for our translators. It will also make for unique
identifiers that are unnecessarily long in most cases.

To avoid this, this commit adds an extra pass over the document to first
find the IDs that are unique. Later when we create the
`TranslatableXmlElement` we check if the ID is unique. If not, we
prepend the GUID. This way we only include the GUID if it is really
necessary.