Exclude framework references from package
When building the net46 version of XliffTasks we need to <Reference> a
number of framework assemblies. With version 2.0 of the SDK NuGet now
takes those elements and turns them into <frameworkAssembly> elements in
the generated .nuspec. This is turn causes any project consuming the
package to *also* get references to those framework assemblies.

This means that consumers of XliffTasks are currently getting references
they don't actually want or need, which may end up conflicting with
references they *do* want. Since XliffTasks is strictly a tool we don't
need this behavior, and can turn it off by adding `Pack="false"`
metadata to the references.