Better reporting of untranslated resources
Currently, when running the `EnsureAllResourcesTranslated` target/task, each untranslated resource ID will be printed out on a separate line. The "blame" also goes to the .targets file containing the target, rather than the original .resx or .vsct. The path to that file appears in the error message.

This commit re-works the error reporting to blame the original file, and joins the untranslated IDs into a single line. This should greatly reduce the output spew in the common case.