Merged PR 138257: Fix ArgumentNullException when checking regex patterns
Fix ArgumentNullException when checking regex patterns.

This repros when scanning a directory which contains files that are not containers (like .nupkgs).

Also, I added a non-zero return code for help output just to be safe that invoking signcheck.exe only returns 0 if signing verification actually run and had no errors.