Changed the TryBuildZipData code to use ZipArchive instead of Package.
This allows ordinary zip files to work but still works for nupkgs and vsixs.

This changed the order of files processed so NestedContainer test expected
output was changed from ContainerOne.dll to ContainerTwo.dll. I'm not sure
if both files should show up since they are the same binary with the same
hash.