Remove project Language restrictions
We will import these props/targets from projects that
aren't c# and VB projects and so we need to remove the
hard error.

For the resource generation we should avoid running the target
if there are no EmbeddedResource items which is an indicaion that
it isn't a C# or VB project.