Add support for some additional parameters to GenerateResxSource
Support a different class name for access vs storing resources.  This is needed to support
.NETNative where we need to handle merging the FX resources into a single assembly but
also wish all framework code to access through internal System.SR classes.

Adds the ability to suppress the implementation of GetResourceString so that it can be
implemented in a separate partial file.

Adds the ability to optionally pass the default value of the resource string.