Do not DelaySign with Open key
Tests expect assemblies to be full-signed as we do not
register the open key for verification skipping.

Previously we'd delay sign these assemblies because
we knew were going to change them with IL-rewriting
before eventually signing via SignTool in the official
builds.

That won't work for testing since tests won't get the
signed output.