Allow for changing the msbuild to be used for the build in arcade
In the powershell script, you can change both the path as well as the command to be used. This was not totally possible in the shell script. This change allows for that to happen.

This is needed at least on msbuild where we want to point to our custom bustom msbuild build to use.