Add additional support to SignTool
This change adds the following:
1) Fixes issue with null list of strong name files to skip verification
2) Adds option to use content hash in extraction path for package extraction
3) Adds option to skip validation of container verification
4) Adds option to read and reuse extracted package caches across runs
5) Adds some additional logging messages to know the files we are working with

This enables someone to run the signtool task more then once and use features
of the signtool task like unpacking and repacking as well as signing validation
while not necessarily using microbuild to sign. Ideally this functionality would
be factored out into separate tasks but for now it accomplishes the job.

For my current team we are using a process along the lines of:
1) Use signtool in DryRun=true mode to extract a set of binaries from packages
2) Use external signing service sign binaries in the cache
3) Use signtool in real mode with the new reuse cache option which
will cause the repackaging and signing validation to run (excluding
package siging validation). Note we also need to dummy out the microbuild
targets to make the msbuild sign job a no-op.
4) Use external signing service to sign packages.