Work around Microsoft/msbuild#3345
`helix.proj` is built with the target `Test`, but does important work
in `Wait`. Since that target was hooked to `Test` by an `AfterTargets`,
it hit a bug in MSBuild that failed `AfterTargets` targets don't fail a
build that builds the target they hook.

Restructure the targets (transparently to callers) to avoid that
situation and observe failures in `Wait`.