Add helper methods for "target" and "state"
Per the XLIFF 1.2 spec, a `trans-unit` element must have a `source` element, but `target` is optional. Similarly, within a `target` element the `state` attribute is optional. As we manipulate .xlf documents we need to keep this in mind. This commit introduces some helper methods to makie it easier to get and set the value of `state` and `target`.

- When getting the `target` value, if the `target` element is not present we instead return the `source` value.
- When getting the `state` value, if the attribute is not present we default to returning "new".
- When setting the `target` value, if the element is not already present we create it, and ensure that it comes immediately after the `source` element as required by the spec (or more precisely, it must come before the `note` element, if any).
- When setting the `state` value, we create the both the `target` element (as above) and `state` attribute if necessary.