Calculate Configuration default value consistently
I missed a place where we advanced the property but didn't recalculate the default value.

As a result, when a config string had two consecutive omitted properties we'd produce a configuration that
was invalid.

Fix this by moving the default calculation into a local method and using it in both places.