Fix ResourceUsageAnalyzer to account for new overloads
This analyzer is trying to prevent calls to SR.Format(SR.Something), which should just be simplified to SR.Something.  I'm adding some new SR overloads to also support culture, and the way the analyzer is currently coded causes it to fire incorrectly.  This just simplifies the analyzer, eliminating the false positives.