Workaround NuGet breaking change
NuGet removed Manifest.Save(Stream, bool) in the latest version.
Since NuGet is part of MSBuild our task is forcibly upgraded to the
latest version when running in the latest SDK.  This results in a
missing method exception.

Workaround it by calling an API that NuGet didn't remove.  This can
potentially cause failures in GenerateNuSpec because it will now run
Validation, however anyone using the Pack task in our package should
already be passing validation since NuGet runs the same validation
when reading the Manifest.