Remove parallelization from these calls
When multiple builds run this code in parallel, the server crashes because it openes
hundreds of tcp connections to our server, which causes it to attempt to open
hundreds of tcp connections to SQL and Kusto.

This always slows everything down, and often results in dropped connections.
We experienced over 1 million failed requests last week because of this.