Ensure MSBuild doesn't downgrade NuGet
MSBuild uses NuGet and lists it in its own deps file.
As a result CoreCLR will unify the version loaded by the task the version loaded in MSBuild.

If the version needed by the task is greater than the one in MSBuild
this results in a FileLoadException.

Workaround this by compiling against the first assembly that has the
LicenseMetadata API needed by e39764c1fcf0664db63aa63c44ea0aca5f9ad36c which is
equal to the version inbox in the SDK 2.1.500.