Make packaging tasks work with PackageLicenseExpression
When using PackageLicenseExpression LicenseUrl must not be defined.
Remove our default which was stale anways.

Workaround an issue in NuGet manifest validation which was rejecting
a nuspec that they wrote themselves.  This can be removed once
we can pick up a new NuGet (which will likely be months since this needs
to make it into MSBuild, then the SDK and we need to adopt that SDK.