Fix NGEN method logging
The NGEN method logging code didn't account for the fact that we run
IBCMerge for the same assembly multiple times when the project uses
multi-targeting. This created race conditions on both writing the temp
XML file and the NGEN method list.

This PR fixes this by properly writing out separate files for each
target framework of an assembly.

closes #2309