Remove STAThread attribute from Xunit.Console.dll
Windows 10 Nano Server does not support using the STA threading apartment. CoreCLR has been recently updated to handle this case and notify the user via an exception (instead of an assert that is ignored in release builds). Because xunit.console.dll has been requiring the STA thread, this has caused our test runs on Windows 10 Nano Server to fail recently.

Since both CoreCLR and CoreFX have been running without issue on xunit.console.dll for a while without having an issue related to failing to initalize the STA thread, I believe that we can remove the STAThread attribute and allow the main thread of xunit.console.dll to run in the MTA.

This will fix all of the Windows 10 Nano Server runs on CoreCLR (debug, checked, and release).

cc: @RussKeldorph  @echesakovMSFT