Add diagnostic information when remote executor times out
When a RemoteExecute.Invoke times out, we get very little detail about what the remote process was doing at the time.  This commit adds details from the Process object, but more importantly, when available attaches ClrMD to the remote process in order to dump out details like threads and stacks.