Make DiffingService public (#3184)
In MEF v1, imports on private properties are fulfilled as as well. However, in
MEF v2 this isn't the case. Sadly, the code handles not having a diffing service
by creating a default one. However, the default one doesn't diff attributes.

This resulted in API Reviewer not reporting differences when they only
occurred in attributes. It seems we haven't noticed this for quite a while...