Add Performance Template (#3175)
* Add Performance Template

Adding the performance testing template.

I have tested this in dotnet/performance (dotnet/performance#582) and dotnet/coreclr (dotnet/coreclr#25390), both of which passed. I am working on getting it prototyped in corefx.

This will not be ready for real use until dotnet/core-eng#6795 and dotnet/core-eng#6815 are complete, at which time we will work with the partner teams to get the template hooked into their builds. At that time, the scripts this template uses will also need to be updated to target the new queues.

* Add extra parameters, fix parameter name

* Fix the WorkItem arguments again

* Address feedback

* Use AGENT_OS, not TargetsWindows

* Add variables parameter

* Use BuildConfig in proj file

* Use $HELIX_WORKITEM_PAYLOAD

* Address additional feedback

* Fix linux core_root_directory

* Actually fix linux

* One more fix for linux

* Address final feedback

* Fix something that broke when I addressed feedback

* Fix some things that were broken

* Use new queues