Add Performance PR Support (#3733)
* Add support for A/B perf runs for prs

This change adds support to perform A/B comparisons of performance testing. This change adds two new options to the performance-setup scripts:

BaselineCoreRootDirectory - The directory where the baseline CoreRun.exe has been built
Compare - a switch to tell the system that we want to compare a baseline to the PR that is being tested.
These two options are not mutually required. A user could supply a baseline, but not the compare switch, but then no comparison would be done. If the user supplies the compare flag, but no baseline, their job will compare against the latest dotnet sdk.

This change also adds the support to the proj file to perform the compare, which involves running the baseline as a precommand, then running the actual test run, then running ResultsComparer as a post command, which has been set up to write out an xml file that matches the xunit format that the xunit reporter expects. For compare runs, we will ignore failures, and pass the job, as deciding if a regression should fail the PR is at the code reviewer's discretion.

This change does not provide advice as to how to create the baseline Core_Root, which is up to the repo owners to decide.

* Add documentation about PR support and motivation for scripts