Error on overlapping content hashes (#3781)
There was a Debug.Assert here for this case, but this issue can be generated by user input. NuGet.Client recently changed to produce packages deterministically, which caused this error to show up in the form of failed signing validation. Change the Debug.Assert to an error so that we can catch this in user repos.
I did not use the LogError(code, file) form of this because it doesn't present a good way to associate the packages that have the errors.