Workaround some bug in AzDO templates (#3807)
The parameter value for continueOnError was not being properly passed through (ending up as an empty value). After some experimentation, this appears to be a bug that can be worked around in two ways:
- Change the caller (post-build.yml) to not have the 'continueOnError' parameter as a nested parameter
- Changing the template itself to not call the parameter name 'continueOnError'.
Note that the "params" nested parameter seems to work just fine. My hunch is that this is becuase the matching template parameter is named different than the caller's parameter.