Don't add dummy dependencies to partial packages
These dependencies were causing issues because they get copied to
newer/compatible frameworks when we live build the newer/more-specific
thing but are missing the other.

We could better plumb exclusions for these dummy dependencies to maintain
the "fail to install" characteristics, but I don't think it's particularly
interesting to try to force the packages to fail install.  If we're missing
the actual DLL the project will fail to compile.