Fix swagger generation (#4079)
On non-nullable, non-required parameter types, the handlebars code will automatically append a ? to the type reference, making it nullable. This logic needs to be duplicated in the null check code, otherwise we end up with:

```
int foo(int? p)
{
    if (p == default(int)) {...}
}
```

This will compile fine but will not have the intended behavior (default(int) is 0 while default(int?) is null).