Add support for pre-relese iterations (#4127)
* Add support for pre-relese iterations
In .NET 5, to avoid issues with preview sorting if there are more than 9 previews, we have decided to take advantage of SemVer2's ability to specify a pre-release iteration number after the prerelease tag.  For example: preview1 -> preview.1.
This commit adds support for this natively in Arcade. Repositories wishing to take part should set PreReleaseVersionIteration to the numeric value of the current iteration (e.g. 1).

Repositories that are not planning on any iteration should leave this property blank, for example:
    - The current label is servicing
    - The repo has alreeady shipped packages to nuget.org with a pre-release iteration built into the pre-release label at the current major.minor.patch (e.g. if they shipped a '5.0.0-beta4' package already)