Don't fail automatically on no artifacts (#4222)
Some repos don't publish packages and blobs (or both) but may want to particpate in publishing for other reasons (symbols for instance). This is prevented by two things:
- The artifact download steps fail if the the artifacts don't exist
- The publishing logic checks that the directories exists.

Instead, pass continue on error to the artifact download steps, and then error out in publishing only if files mentioned in the manifest are missing.