Add functionality to supply artifact patterns in execute-sdl.yml (#4306)
Some repos upload a large number of artifacts that are not needed for sdl validation. This change allows users to specify various patterns for artifacts to download. If none are supplied, we default to downloading all artifacts.

This change adds the functionality to execute-sdl.yml to allow the user to supply a list of artifact folders to download.

Usage:
- template: /eng/common/templates/job/execute-sdl.yml
      parameters:
        artifactNames: 
          - 'Folder1'
          - 'Folder2'