Open xml files with utf-8 encoding (#4349)
In the xunit-reporter, we try to open xml files that are encoded in utf-8. This worked fine in python2, because python2 assumed file encodings differently. Python3 tries to use whatever the "default" file encoding for the machine is, which is not utf-8 on windows, while our xml files with test results are encoded using utf-8. This change explicitly adds the utf-8 encoding to the file open call.