Include last modified time in package index cache key (#4012) (#4367)
* Include last modified time in package index cache key

We were seeing cases when stabalizing packages that the index may have
already been cached in the MSBuild process, resulting in some projects
seeing the update and others not.  Avoid this by including modified time
in the cache key.  If the file is changed it will be a cache miss.

* Include file length in packageIndex cache key