Fix Alpine rootfs build
Specifying the edge/main and edge/testing repos for all the
packages lead to an unintended installation of libstdc++.so.6.0.27.
The libstdc++.so.6.0.25 should be installed instead, otherwise
the binaries we build have unresolved dependency to
`std::__cxx11::basic_stringstream<char, std::char_traits<char>, std::allocator<char> >::basic_stringstream()`
that got apparently added in the latest version of the library.

This change fixes it by separating installations of packages from
the edge/testing and edge/main.