Use .txt instead of .html for file attachment in AzDO test runs
Right now to view the console.log of a Helix test run you need to download the UploadFileResults.html, open it in your browser, click and download the console.log and finally open it.

AzDO renders a preview of .txt files so we can save downloading the .html file by switching to .txt.

Fixes https://github.com/dotnet/arcade/issues/4695