Support TargetFramework.Sdk in SymStore.targets
Fixes https://github.com/dotnet/runtime/issues/32300

This fixes the frequent failures where multiple inner builds try to write to the same location in combination with SymStore.targets. As discussed in the linked issue, we should investigate if we can change the TargetFramework.Sdk to not depend on such hacks.