Support GenFacadesForceZeroVersionSeeds for source-based facades
When moving to source-generated facades we didn't initially support zero
versioned seeds.  This turned out to be problematic in servicing, since
our facades have dangling references to assemblies which ship in
packages outside the framework.  Those references would update when the
package was serviced, an application may see the updated shims when
a new shared framework was installed, but not necessarily see the
package update since that would require an app rebuild.

To fix this we're implementing zero-versioned seeds for shims by
rewriting the assembly after compile.