Optprof download diagnosability (#3974)
* Specify optprof prefix early

In cases where the prefix is computed incorrectly, an official build failure
is hard to diagnose because the console messages and warnings did not contain
the queried path. Move the messages to an earlier point.

* Distinguish OptProf download errors

Create distinct errors for "no drops were found at all" and "no valid drops found".