Skip component detection and codesign validation for some jobs (#5049)
We've had an increase in the time for a bunch of standard build tasks
(publishing, validation, and some parts of the build stages). Looking at
some builds seems to show that a lot (all?) of this increase comes from
injected tasks. We don't need these for the following standard templates
because they create no actual content:
- Publish to Build Asset Registry
- All of post build (sdl validation and publishing)
Skip these injected tasks in these cases.