Reduce aka.ms API batch size and improve efficiency (#5067)
* Reduce aka.ms API batch size and improve efficiency
Reduce the batch size to the aks.ms API. We were hitting the maximum
content length in some cases. The content length is 50k, but the maximum
links is 300.
Also change around the way we use the API to first attempt to update the
links, then only if bad request is returned, bucketize and create+update.
This should drastically reduce the total time required for generation.
That said, the time is still not great. A batch of 100 takes ~25 seconds,
so additional optimizations will be coming soon, like making the whole
process more async

* Update src/Microsoft.DotNet.Deployment.Tasks.Links/src/AkaMSLinksManager.cs

Co-Authored-By: Ricardo Arenas <riarenas@microsoft.com>

* Respond to feedback

Co-authored-by: Ricardo Arenas <riarenas@microsoft.com>