Add MSBuildRestoreSessionId property to restore
MSBuild usually sets this property to invalidate the cache but that only happens during evaluation of the invoked project. Invoking restore on projects explicitly means that we need to set that property manually. Setting this is general goodness, as we can condition on it so that we know that we are inside the restore phase.