Symbol validation continue on error (#5181)
* Add ContinueOnError to symbols-validation

For the release pipeline, we would like to validate all of the symbols. Currently, we exit as soon as we find one package with missing symbols, which means that we would have to fix the problem for each pacakge sequentially, starting over with the entire pipeline. This change adds an optional parameter, ContinueOnError, that will allow us to note the failed packages individually, and continue to check all of the other packages, failing once we have checked all of them.

* Fix indentation