Support VS single evaluation mode
With the recent change to support RestoreUseStaticGraphEvaluation, the hook point before Restore was removed as it was assumed that the TargetFrameworks modification can be done during evaluation. As VS shares the same evaluation for restore and build, the modification is ignored (as it is conditioned on MSBuildRestoreSessionId and IsGraphBuild). Changing the condition to only check for MSBuildRestoreSessionId or building inside Visual Studio (as VS doesn't set MSBuildRestoreSessionId).