Workaround NuGet\Home#9532
A regression in JsonRuntimeFormat.WriteRuntimeGraph caused it to omit
the root object when writing out the runtime graph.

Workaround this by copying WriteRuntimeGraph(string,RuntimeGraph)
and calling StartObject/EndObject.

Add a test to validate round tripping.  Test should fail when bug is fixed
and we can remove the workaround (and keep the test).