Fix TestPlatformApplies when Any is used.
If the method receives a TestPlatform.Any flag it does not need to test
against all the platforms, since they are all included. Simply check for
that value first to make sure that we will return true.

This was discovered when trying to get the Traits on iOS and Android in
the ActiveIssue attribute which was returning none.