Fix sourcelink and symbols validation (#5589)
* Fix sourcelink and symbols validation

While testing adding sourcelink to the release pipeline, I noticed that we were not reporting the correct number of failures. Noteably, there would be packages with broken sourcelink links (as noted by the output text), but we'd report 0 failures. After some investigation, I discovered that that is because once a job is received using Receive-Job, all of its outputs and results are discarded. Additionally, just because a scriptblock returns non-zero, doesn't mean it comes up as "failed." It is reported as "completed." Therefore, when we would receive all of the completed jobs, and then  remove them, we would lose their job result, and not count them as part of the total failures, making it seem like we had fewer sourcelink failures than we did.

The solution for this is to count the failures as we receive the completed jobs. This change fixes that.

* fix typo