Permit inclusion of build-time binaries in pkgproj
PkgProj assumes that every dll that goes into a package is either a
reference or implementation assembly for with the same name
as the package.

The same-name requirement was a bad one, this just gets in the way when
we need to suppress it.  The requirement is not imposed by NuGet Pack
target.

We were also examining all DLLs in the package for assembly references
to raise to PackageReferences, but this should only be done for
ref/lib/runtime assets.