Factor installer light package task into a separate arcade task package (#5806)
* Factor installer light package task into a separate arcade task package
This moves CreateLightCommandPackageDrop into a separate arcade task package,
so that the functionality can be reused in some other repos that are not always
building shared framework assemblies.

Ideally this package will eventually factor out all installer task functionality
so it can be shared among all repos. The issue right now is that the way wix installers
are built in the runtime repos, aspnetcore, and installer are all very different.
Installer uses powershell scripts to invoke wix, aspnetcore uses the wix nuget package,
and the runtime repos use the shared framework sdk, which invokes wix using Exec.

I think the right way to do this long term is to use the wix nuget package. This would mean
that this package would reference a version of wix, exposing the tasks, and add its
additional tasks on top of that. Similar functionality could be built for other types of installers.

* Add missing endproject