Fix int nupkg id calc: use SourceBuildRepoName
The intent is that SourceBuildRepoName gets added to an existing element so the source-build intermediate nupkg comes automatically with updates to that existing dependency. Change the Version.Details.xml task to return that value as the identity, and the targets to use it that way.

Also rearrange the targets so the logic to calculate intermediate nupkg IDs is shared between the restore and production side. This change aligns the source-build targets files more closely with the projects that import them.