Remove the validate-darc call
Darc requires 3.1 to run (rather than the 3.0 installed by the global.json)
However, there is a breaking change in the 3.1 NuGet packages
that causes Sleet to fail.

This is a change to unblock the builds by going back to 3.0 (which
is the change that caused this break).

There is a longer term plan to make the "right" change here,
which is to remove Sleet altogether.