Update destination for dotnet-cli to avoid collisions
In dotnet/runtime we submit our own dotnet host which is used to invoke VSTest's testhost. Still we require a full sdk to invoke dotnet test (which invokes vstest.console). When settings IncludeDotNetCli to true, we get a collision as our dotnet file conflicts with the dotnet folder that is bootstrapped by the Helix SDK. Changing the helix destination for dotnet-cli to dotnet-cli to avoid such collisions.