Correct the version of xcopy msbuild. (#6194)
* Correct the version of xcopy msbuild.
The min VS version cannot be conveyed with preview tags, so use 16.8 instead. Also fixup a place in here were there was an assumption that the xcopy msbuild would always end in 0.-alpha. This is an incorrect assumption, and in hard to identify code. Instead, pull out the default package version next to the min vs version and rework the code so that it doesn't assume suffixes. To use preview versions that aren't the default, repos should specify the xcopy msbuild specifically in global.json

I don't think this will break anyone. Because the min default is changing, in order to hit the code path that assumes a 16.8.0 package, the user would have had to specify a min VS version in their global json that has a major.minor that is newer than 16.8. This doesn't exist yet, so we shouldn't hit this path.

* Update eng/common/tools.ps1

Co-authored-by: Ricardo Arenas <riarenas@microsoft.com>

* Remove extra logging

Co-authored-by: Ricardo Arenas <riarenas@microsoft.com>