Include the filename in the zip data map key (#6317)
* Include the filename in the zip data map key
.NET has a bunch of cases where symbol packages and non-symbol packages are identical.
This occurs when there are either no symbols in a symbol package (e.g. a sources package), and
cases where an original package also contains symbols. SignTool does not like this.

The primary issue is that the zip data map (for each container) only uses the content
hash as the key. This means that zip files with different file names will have the same content
hash and map to the same zip data. This is problematic because signing information can
be **different** based on filename. For example, we may choose to sign .nupkgs but not
the corresponding .symbol.nupkg. If the content hash is the key for the zip data map,
then the .symbol.nupkg and .nupkg are de-duplicated and the file signing information we choose to use
is arbitrary.

To fix this, use the file name in the zip map key, like we do file the map of files to their signed data.

* Add test for hash conflict with different file names

* Keep collision priority id when dealing with ignored file extensions

* Avoid wixlib checks