Add retry capability to symbols-validation
When running on a large number of symbols packages, we can occasionally get 503 errors. This doesn't mean that the symbols don't exist on SymWeb/MSDL, only that when trying to access them, something went wrong. In that case, retry. Testing suggested that 3 attempts was plenty.

This change also fixes up some returns from CountMissingSymbols, which hadn't updated to the new model.