Fix MSBuild escaping issues (#6510)
* Fix MSBuild escaping issues

MSBuild does some weird escaping of strings in metadata when we use the basic constructor (ie, it removes any escaping behavior/assumes % as an escape character and compresses \\). We do not want that behavior. We want all characters to be treated as the literal character. This change modifies how we save the metadata by using ITaskItem2 and SetMetadataValueLiteral.

Fixes #6163 and #6490.

* Address PR feedback

* Change to IDictionary