Update arcade (#6609)
Update arcade to unblock publishing for cases where places have taken a new arcade.
The new arcade doesn't generate certain attributes that are not required in the signing info node in the manifest, but the arcade currently in use in arcade still expects them. This arcade is what is used for publishing, so it's not possible to publish an arcade with altered signing attributes without updating arcade with this arcade.