Fixup item to sign check and associated path (#6613)
When creating the signing model, we check that an item to sign appears in the artifacts. The check for a package was incorrectly constructing the package file name. In arcade, there are various hard rules that a package Foo with version 1.0.0 must appear as Foo.1.0.0.nupkg.
Rename the test input package to have this form, and update the test that was verifying this case to be stronger and correct according to the arcade package model.