Add source index job to jobs.yml (#6741)
This change adds an optional source indexing job to the pipeline so that the indexer won't just magically break when the repo gets changed. The job builds and collects the required files for indexing, and uploads them so the indexer can run without large dependencies on how individual repositories build.