Publish using Darc should depend on the publishDependsOn (#6770)
We specify publishDependsOn in the parameters, but hard-coded Validate for publish_using_darc. What this means is that, in particular for arcade, where we Validate the arcade sdk in addition to the normal validation legs, we don't wait for ValidateSdk to finish before publishing using darc. However, we should be waiting for that, so we should be using the publishDependsOn, rather than hard-coding Validate for publish using darc's dependsOn.