Exclude Newtonsoft.Json dependency from pkg
On .NETCoreApp, exclude the Newtonsoft.Json dependency in Packaging
package as the SDK already comes with its own copy which is being loaded
into its own ALC. If the Newtonsoft.Json dependency is present in the
package, unification during runtime won't work as the two Newtonsoft
assemblies are loaded into different ALCs. This would then result in a
MissingMethodException when trying to use exchange types from
Newtonsoft.