Fix failing unit tests in Arcade SDK and include them in CI test runs (#7159)
This commit re-enables `Microsoft.DotNet.Arcade.Sdk.Tests` project when during Helix test runs.
Previously it was disabled because there were some problems with test environment on Helix.

Additionally it fixes the following test failures:

### FindLatestDropTests

`JObject.Parse` parses JSON dates as `DateTime` by default not as `string`. When we were calling
`(string)item["CreatedDateUtc"]` the `DateTime` stored in that property was converted to string
using invariant culture and then we tried to parse it back using current culture. It worked correctly
as long as the date time formats for both cultures were the same but this is only true for `en-US`.

I also converted `GetLatestDropName_Error` test to use `Theory` instead of having four asserts inside.
This makes it easier to see which case actually failed.

### GenerateResxSourceTests

The generation code was changed some time ago but the test data that we use as the expected output
was not updated. I copied the current output of the generator and put it into test assets. Of course
this assumes that the current output is correct but it looks like it is.

### GenerateSourcePackageSourceLinkTargetsFileTests

The tests failed on Unix because some of the backslashes in test strings were converted to front slashes
but not all. It was fixed by explicitly normalizing all paths before passing them to the tested code.

### MinimalRepoTests / RepoWithConditionalProjectsToBuildTests

I had to mark these tests as skipped because they are not trivial to fix and they block us
from enabling Arcade.SDK.Tests on CI builds. We should enable them back as soon as they are fixed.

Related issue: https://github.com/dotnet/arcade/issues/7092

### To double check:

* [X] The right tests are in and and the right validation has happened.  Guidance:  https://github.com/dotnet/core-eng/tree/master/Documentation/Validation