Add more retries to symbols and sourcelink validation
For sourcelink validation, this change introduces retries.

For symbols validation, this change expands retry scenarios to any failure, not just 503s, as sometimes, dotnet-symbol fails with no real indication why, and we want to retry on all scenarios. Also updates the logging, and separates out the files being downloaded via dotnet-symbol to separate directories based on msdl/symweb, so we don't think we have those files on both when they only exist on one, as our method for confirming if they were downloaded just checked to see if certain paths exist.