Make ModifierComparer use type references instead of definition
When comparing definitions this was failing if the type couldn't be resolved,
for instance if a type-forwarded dependency was missing or references
were specified at all.

This was inconsistent with our other parameter comparisons which just
evaluate the compare the type reference as a string without considering
assembly.

This does raise a concerning issue, that APICompat doesn't actually
evaluate assembly equivalence of parameters but that's an existing
problem.