Recognize files in ref-pack under "analyzers" as analyzers and write them to the frameworkList (#7561)
* Recognize files in ref-pack under "analyzers" as analyzers and write them to the frameworkList

* Fixes to support analyzer packaging

Platform manifest creation uses a dictionary to map template items to
files in the package, but didn't account for duplicate file names (as is
the case with .resources.dll).  Handle this by taking the first file of
a particular name.

Include `analyzer/` prefixed files in framework list.

Allow for analyzers to be excluded from closure and type validation with
ExcludeFromValidation metadata.  This metadata is necessary because
analyzers will be missing dependencies provided by the compiler, and
their resource.dll will have duplicate files, both which would result
in errors.