Publish log folder without configuration
In https://dev.azure.com/dnceng/internal/_build/results?buildId=1250178&view=results we hit an issue where the "Restore internal tools" step failed, but AzDO only highlights the failure in the "Publish Logs" step because the directory didn't exist.
This happens because the restore step writes a file in `D:\workspace\_work\1\s\artifacts\log\Debug\` but the publish logs step only looks for the `Release` directory and fails if it doesn't exist.

Publish the whole log folder instead, which is what we do in other similar steps anyway.